+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add "set-log-level" subcommand for controller and node
+
 ## [1.20.0] - 2023-10-11
 
 ### Added